+++ public/js/main.js
@@ -3309,7 +3309,7 @@ CategoryModule.factory('Category', CategoryService);
 // Category module controllers
 CategoryModule.controller('CategoryListController', CategoryListController);
 
-var ReaderViewController = function($scope, $rootScope, $http, $timeout, Post, Upload) {
+var ReaderViewController = function($scope, $rootScope, $http, $timeout, Post, Upload, modalService) {
 
   $scope.post = {};
   $scope.comment = {content:''};
@@ -3508,14 +3508,23 @@ var ReaderViewController = function($scope, $rootScope, $http, $timeout, Post, U
 
   // Comment deletion
   $scope.deleteComment = function(comment) {
-    $http.delete(layer_path + 'post/comment/' + $scope.post.id + '/' + comment.position)
-    .then(function() {
-      var position = $scope.post.comments.set.indexOf(comment);
-      if(position > -1) {
-        $scope.post.comments.set.splice(position, 1);
-        $scope.post.comments.count--;
-      }
-      $scope.$broadcast('scrubberRecalculate');
+    var modalOptions = {
+      closeButtonText: 'Cancelar',
+      actionButtonText: 'Eliminar comentario',
+      headerText: '¿Eliminar comentario?',
+      bodyText: 'Una vez que se elimine, no podrás recuperarlo.'
+    };
+
+    modalService.showModal({}, modalOptions).then(function (result) {
+      $http.delete(layer_path + 'post/comment/' + $scope.post.id + '/' + comment.position)
+        .then(function() {
+          var position = $scope.post.comments.set.indexOf(comment);
+          if(position > -1) {
+            $scope.post.comments.set.splice(position, 1);
+            $scope.post.comments.count--;
+          }
+          $scope.$broadcast('scrubberRecalculate');
+        });
     });
   }
 
@@ -4812,6 +4821,55 @@ boardApplication.controller('MainController', ['$scope', '$rootScope', '$http', 
       error(function(data) {});
   }
 ]);
+
+boardApplication.service('modalService', ['$modal', function ($modal) {
+  var modalDefaults = {
+    backdrop: true,
+    keyboard: true,
+    modalFade: true,
+    size: 'sm',
+    templateUrl: '/js/partials/modal.html'
+  };
+
+  var modalOptions = {
+    closeButtonText: 'Cerrar',
+    actionButtonText: 'Aceptar',
+    headerText: '¿Estás seguro?',
+    bodyText: '¿Quieres realizar esta acción?'
+  };
+
+  this.showModal = function (customModalDefaults, customModalOptions) {
+    if (!customModalDefaults) customModalDefaults = {};
+    customModalDefaults.backdrop = 'static';
+    return this.show(customModalDefaults, customModalOptions);
+  };
+
+  this.show = function (customModalDefaults, customModalOptions) {
+    //Create temp objects to work with since we're in a singleton service
+    var tempModalDefaults = {};
+    var tempModalOptions = {};
+
+    //Map angular-ui modal custom defaults to modal defaults defined in service
+    angular.extend(tempModalDefaults, modalDefaults, customModalDefaults);
+
+    //Map modal.html $scope custom properties to defaults defined in service
+    angular.extend(tempModalOptions, modalOptions, customModalOptions);
+
+    if (!tempModalDefaults.controller) {
+      tempModalDefaults.controller = function ($scope, $modalInstance) {
+        $scope.modalOptions = tempModalOptions;
+        $scope.modalOptions.ok = function (result) {
+          $modalInstance.close(result);
+        };
+        $scope.modalOptions.close = function (result) {
+          $modalInstance.dismiss('cancel');
+        };
+      }
+    }
+
+    return $modal.open(tempModalDefaults).result;
+  };
+}]);
 
 boardApplication.run(['$rootScope', '$http', 'AclService', function($rootScope, $http, AclService) {
   // TEST PURPOSES
